Add relative path support and refactor
