+++ package.json
@@ -25,7 +25,7 @@
     "jest": "27.5.0"
   },
   "dependencies": {
-    "lodash": "4.17.21",
-    "@kaholo/plugin-library": "^1.3.0"
+    "@kaholo/plugin-library": "^2.1.2",
+    "lodash": "4.17.21"
   }
 }
